[skip ci] Add a check to unfeasible analyzer : renaming files + move diagnostic interface to another source file
