+++ src/solver/infeasible-problem-analysis/unfeability-diagnostic.h
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace Antares::Optimization
+{
+class InfeasibilityDiagnostic
+{
+public:
+    InfeasibilityDiagnostic(std::shared_ptr<operations_research::MPSolver> problem_);
+    virtual void run() = 0;
+    virtual void printReport() = 0;
+    virtual std::string title() = 0;
+    bool hasDetectedInfeasibilityCause() { return hasDetectedInfeasibilityCause_; }
+
+protected:
+    std::shared_ptr<operations_research::MPSolver> problem_;
+    bool hasDetectedInfeasibilityCause_ = false;
+};
+}
